requirements.txt
# Core RAG and LangChain dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0

# Vector database
chromadb>=0.4.0

# Document processing
pypdf>=4.0.0
python-docx>=1.0.0
unstructured>=0.11.0,<0.12.0

# Embeddings and text processing
sentence-transformers>=2.6.0
tiktoken>=0.6.0

# Web framework
streamlit>=1.30.0
streamlit-chat>=0.1.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0

# Optional: For better performance
faiss-cpu>=1.7.0

# Advanced analytics and intelligence
nltk>=3.8
scikit-learn>=1.3.0
networkx>=3.0
matplotlib>=3.7.0

# Offline ML models for no-API functionality
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0